requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
django-filter==23.3
djoser==2.2.2
gunicorn==21.2.0
celery==5.2.7
redis==4.5.1
flower==1.2.0